A new script for spatial LookerUp.
The spatial_utils script inherents from
axelrod_utils. The only method that changes is that of the scoring.
THe score of a strategy(table) is not based on the average score
of a round robin tournament anymore, but it is based on the rank of that
strategy on a series of spatial tournaments.

do_table and score_tables are alterd on their arguments.