Reading in the previous results.
If we want to re run the genetic algorithm, from the point the
previous one ended, we will use the previous output.

The evolve function outputs the progress of the genetic algorithm.
Thus by using a file exist command, if the algorithm has been previously
perfomed we are using the output.

Once the output is reade in, we are re-id the output. The key are
turned into a full table.

Also changed the default output name of the file.