Spatial utils.
This script contains the evaluation functions used to
score the tables. That is the median or max rank divided
by the tournament size.

We have made sure that at least one tournament happens. And that
we are generating graphs (3) for the same given players and parameter
p.

SOme long code has been broken into smaller functions. Such as :
- topology, returns that graph based on the topology passed as an
  argument.
- ranking, plays the spatial tournament and returns the rank of
  LookerUp.